ai_agents/strategy_agent/requirements.txt
# requirements.txt pour strategy_agent

# LangGraph et LangChain
langgraph>=0.0.55
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0

# Modèles de données et validation
pydantic>=2.0.0
typing-extensions>=4.0.0

# Utilitaires
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.24.0

# Logging et monitoring
structlog>=23.0.0
prometheus-client>=0.16.0

# Tests
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Sérialisation et données
orjson>=3.8.0
pandas>=2.0.0

# Configuration
pyyaml>=6.0
toml>=0.10.2

# Optionnel pour l'interface
streamlit>=1.28.0
gradio>=4.0.0

# Base Python
python-dateutil>=2.8.0
uuid>=1.30